flutter: Fix gradle/wrapper/gradle-wrapper.properties is missing distributionSha256Sum.
This means that the gradle download is not verified.
We recommend explicitly setting the expected Sha256sum to protect you and your apps if a bad actor gets access to the Gradle servers or manages to MitM your internet connection.
Please note that Android Studio does not currently support this and may or may  not crash, but we suggest you try our suggestion anyway to improve the security of yourself and your app.

See: https://gitlab.com/fdroid/fdroiddata/-/merge_requests/8917