Delete src/ui/flutter_app/ios directory
This App cannot release for iOS.

Because iOS Developer Program License AgreementPurpose:

If Your Application includes any FOSS, You agree to comply with all applicable FOSSlicensing terms. You also agree not to use any FOSS in the development of Your Application in such away that would cause the non-FOSS portions of the Apple Software to be subject to any FOSS licensingterms or obligations.