Support importing PlayOK notation
At PlayOK (until April 4th, 2008 it was called Kurnik) a different notation is used,
the fields are numbered line by line in groups of 3 from top to bottom.
So on the top line are the fields 1-2-3, then follow 4-5-6 until finally the last line
is determined by 22-23-24. This notation is much easier to learn than the official one
suggested by the WMD.