RTL: Fix animation icon is too high when locale is not Farsi
This fix was originally intended for Arabic.
I'm not sure if this problem exists in other RTL languages other than Persian and Arabic.
Because there is no translation of other RLT languages, so I modified it first.

This is the refinement of commit 5bf98541300dc9590b8684686ef7df0128be6c15.