Fix CID 338449: Uninitialized scalar field (UNINIT_CTOR)
uninit_member: Non-static class member callsCnt is not initialized in this constructor nor in any functions that it calls.
uninit_member: Non-static class member stopOnPonderhit is not initialized in this constructor nor in any functions that it calls.
uninit_member: Non-static class member ponder._M_base is not initialized in this constructor nor in any functions that it calls.