cleanup variables
Use .last to get the last value of a list

This is less perforamtn as .last iterates over every element in the list to get the last one.
As lists are consecutive itterables it is more efficient to access them directly by thier index (just as it was before).

Our lists aren't that big and we might get away with it. at least for now switching to this style as it is way more readable. We'll see...

fix