Change top-left AnimatedIcon color from black to white (Fix #493)
The color is the same as the status bar of the phone.

Some users may set the background to black, at which point
the AnimatedIcon is invisible. And almost no user selects
the background to white, so setting the AnimatedIcon to white
can ensure that it can be seen in most cases.