Fix clicking dialog option 'Move list' will throw an exception (WAR)
Originally, every time ScaffoldMessenger.of(context).clearSnackBars(); is called,
it will check whether DB().generalSettings.screenReaderSupport is enabled,
which may be missed when refactoring the code before.

See https://sourcegraph.com/github.com/calcitem/Sanmill@master/-/blob/src/ui/flutter_app/lib/widgets/game_page.dart

However, even with this modification,
when DB().generalSettings.screenReaderSupport is enabled,
when ScaffoldMessenger is not,
there is still a problem.
So here is just a temporary solution.

It also needs to be processed when importing move list, which is not processed yet.