Fix #575: Fix the judgment failure of draw
Use other way to Fix:
When just entering the moving phase, when selecting,
history records the record of the last placing.

The problem was introduced in bbea028cc2a474f244d58dedcebaba31e5265032.
After setting _record to null, where it is judged that _record is null,
and it is no longer added to posKeyHistory.

See:

if (MillController().position._record != null &&
          MillController().position._record!.move.length > "-(1,2)".length) {
        if (posKeyHistory.isEmpty ||
            posKeyHistory.last != MillController().position.st.key) {
          posKeyHistory.add(MillController().position.st.key);
          if (DB().ruleSettings.threefoldRepetitionRule &&
              MillController().position._hasGameCycle) {
            MillController().position._setGameOver(
                  PieceColor.draw,
                  GameOverReason.drawThreefoldRepetition,
                );
          }
        }

Change-Id: I2b0fa71ae551b2434937ef8d71600c891a1108f5