Fix 434497 Result is not floating-point
The result of the division is truncated to an integer (a whole number), which is usually a loss of precision in a calculation.

In BoardItem::​pos2polar(QPointF, File &, Rank &): When dividing two values of integer types, integer division is used, which ignores any remainder. When such a result is used in a context expecting a floating-point number, it is likely that floating-point division was intended.

Change-Id: I1af71ca350c15e29112ab0cc54f85fbd3e0a5fab