Refactor game.dart
In this refactored version, we created a new `Player` class
and encapsulated the logic and state related to the game mode
into this class. We also set the `gameMode` via the constructor.
Finally, we reuse some logic in the `Game` class to avoid code
duplication. This makes the code more maintainable and readable.

Change-Id: I6bee56a2bb84b1a10c89f2c93ca6fa8b3592f00c