movepick: Attempt to block opponent's mill in flying phase
Fixed sorting issue, initially, I assumed that for 9mm, when the "to" square
is an odd number, and there are already 2 opponent's pieces adjacent,
I can't move to that point. However, I overlooked the fact that if in the
flying phase, you can jump to that point to stop the opponent. And if the
flying rule is not enabled, generating this move would be impossible.

This modification helps enhance the efficiency of alpha-beta pruning
during the flying phase.

Case:

Case:

 1.    d2    b4
 2.    d1    f4
 3.    d3xb4    b4
 4.    d6    d7
 5.    a4    c4
 6.    e4    e3
 7.    g4    f6
 8.    f2    b2
 9.    a7    b6xd6
10.    d1-g1    d7-d6xa7
11.    g1-d1xf4    f6-f4
12.    g4-g1    f4-f6xg1
13.    f2-f4    d6-d5
14.    d1-a1    d5-d6xa1
15.    a4-a1    c4-c5
16.    a1-d1xc5    b4-c4
17.    d1-a1    c4-b4xd2
18.    a1-a4    d6-d5
19.    d3-d2    b4-c4
20.    a4-b4    d5-d6xb4
21.    d2-g4xc4    b6-b4

Before the modification, White did not prioritize blocking the opponent's mill.
This issue becomes more apparent when MCTS is enabled.

Change-Id: Icb638ef0bcd726b42abe84d4d15ae7ed85933ead