mcts: Add SEARCH_PRUNING_FACTOR but not enable
The SEARCH_PRUNING_FACTOR is used to reduce the number of moves
considered during the search process. A larger value results in
fewer moves being considered, potentially speeding up the search
at the cost of search accuracy. The value should be adjusted
according to the specific problem and performance requirements.

When the value of SEARCH_PRUNING_FACTOR is increased, or other parameters
are adjusted, it may cause stack overflow. In this case, it is possible
that the depth of the tree increases rapidly, resulting in a stack overflow.

The solution is to change the recursive delete to iterative delete.
In this approach, a stack or queue is used to keep track of which nodes to remove,
rather than relying on the system stack. This implementation can solve the problem
of stack overflow, even if the depth of the tree is very large.

Note that this causes additional memory to be allocated on the heap.
In most cases, this shouldn't be a problem.

Change-Id: Id8b5b6c5f52f8b683c765574a983f9106256951d