Refactor to Use View.of(context) for Physical Size
This commit replaces all instances of
`PlatformDispatcher.instance.views.first.physicalSize` with
`View.of(context).platformDispatcher.views.first.physicalSize`.

The change was made to ensure the use of the correct view, particularly in the
context of multi-view applications. By using `View.of(context)`, we ensure that
we're operating on the view related to the current BuildContext, rather than
just the first view that was created.

This provides a more reliable method of accessing the physical size of the
relevant view in a multi-view application. Please note that for single-view
applications, both methods should provide equivalent functionality.

Change-Id: I9324fcc3532507fa859b8bc3dc7173c49bfafd6e