perfect: Introduce the Gabor Malom API
We would like to express our sincere gratitude to
Gabor E. Gevay and Gabor Danner for providing the MalomAPI, a VB.Net dll file.
This resource has significantly facilitated our access to the perfect databases,
and their support has been instrumental to our project.

We appreciate the efforts of these experts in developing this API and for their
continued support throughout our work. Their dedication to their craft is truly
inspiring, and we are fortunate to have had the opportunity to collaborate and
learn from their expertise.

We sincerely thank Gabor E. Gevay and Gabor Danner for their invaluable
contribution.

Change-Id: I474f6d0b9315dfce4e42516c21e918911625f879