Fix sometimes clicking the Move List button does not respond
This problem is currently sometimes reproducible on Windows, but not on Android.

Add delay to fix _MoveListDialog display issue after closing _MoveOptionsModal

In this commit, we added a 100ms delay using `Future.delayed()` to ensure that
the `_MoveListDialog` is displayed correctly after closing the `_MoveOptionsModal`.
This delay prevents the `_MoveListDialog` from disappearing and allows the users
to see it as intended. The context passed to `showDialog` is now valid,
and the semi-transparent background effect is preserved.

(Bug introduced by commit a4e9d58d1520d1cc1a1357afa748cce6509e7681.)