Append current locale to helpImproveTranslateURL
When a user clicks on the 'Help Improve Translate' link, the locale of the
user's device is now appended to the end of the URL. This ensures that the
user is directed to the translation page that matches their device's
language setting. The locale is retrieved using Flutter's Localizations class.

Change-Id: Ia2b7355aa5c6a825a8d2d5567d8bdf7134356a27