Fix sscanf format string to prevent buffer overflow
The previous version of the sscanf call in Position::command had
an unspecified string length, potentially leading to buffer
overflow. This commit modifies the sscanf format string to specify
a maximum string length of 63 characters for moveStr, which has a
total buffer size of 64 bytes. This ensures that the string is
properly null-terminated and prevents buffer overflow
vulnerabilities.

Change-Id: I07a6bd706a53a9768842b8d2190dc5a346064f97