qt: Fix the clock is counting down (WAR)
The function you are referring to seems to be calculating and updating the
remaining time for each player in a game. From the code, it appears that
`timeLimit` is the total amount of time each player has for their moves.
The `remainingTime[WHITE]` and `remainingTime[BLACK]` variables are
updated in the `timerEvent` function. The remaining time for each player
is calculated as the original time limit minus the time already elapsed.

The `elapsedSeconds` for each player is updated in the `updateTime` function,
 which is called from `timerEvent`. The current time is fetched and subtracted
from the `startTime` (which is set when the game starts) and the time elapsed
 by the other player.

In this context, the `timerEvent` function appears to be a callback that gets
 called every 100 milliseconds (as set by `timeID = startTimer(100);` in the
`gameStart` function). This function updates the elapsed and remaining time
 for the players.

If `timeLimit` is greater than 0, the time is treated as countdown time (i.e.,
 time remaining for each player to finish their moves). If `timeLimit` is 0
or less, `remainingTime[WHITE]` and `remainingTime[BLACK]` represents the
time used by each player.

This is why, as you mentioned, `time1` behaves like a countdown timer rather
than an upward counting timer.

Change-Id: I917b1ea045e7a8452dff7bf41fc4ceb83be2866b