qt: Fix inconsistent click behavior on game pieces
This commit addresses an issue where clicking near the edge of a game piece
resulted in no action. The problem was introduced in commit
d0018f2ac25eedcc903f242ed6b44545835fbf66.

Changes:

- Modify the signature of `handlePieceClick` to accept
  `const QGraphicsItem *item` instead of `QGraphicsSceneMouseEvent *mouseEvent`.
- Use `item->scenePos()` for emitting the `mouseReleased` signal to ensure the
  actual position of the game piece is considered, rather than the scene position
  of the mouse event.

By aligning the method used to determine the position of the game piece,
the click behavior is now consistent and accurate.

Change-Id: I6c79562446a89193a2e2a721df3044e0e11f430a