cpp: Change variable type to ptrdiff_t to avoid data loss warning
The 'count' variable type was changed from 'int' to 'ptrdiff_t'
in the Position::has_game_cycle() function. This was done to avoid
a potential data loss warning which was occurring due to the conversion
from '__int64' to 'int'. Now, the 'count' variable can safely hold
the result of the std::count function without any risk of data loss.

Change-Id: If70c86c88c62a7b5837a9c36f5ee7e3e29e4efb4