l10n: sq: Change App name to Mulli
In Albania, the game Nine Men's Morris is more likely to be called "Mulli",
a word that means "mill" in Albanian, in keeping with the theme of the game.
"Mulliri" is usually used to refer to multiple mills or as the name of a
specific mill, whereas in the context of this game it is more likely to
be simply "Mulli".

It is important to note that this game may not be particularly popular
in Albania, so it may not have a widely recognized Albanian name.

Change-Id: I147d4915ef91f0292bd3690186ab73d35ef4fbfa