perfect: Be stronger if skill level more than 15
If the skill level is greater than 15, then the range of move choices for
the perfect database is narrower, aiming to end the game with an equal
number of pieces for both players. If the skill level is less than 15,
although the AI's move choices also result in a draw or a win, it does
not pursue having more remaining pieces than the opponent.

The final effect should be based on the conclusions discussed in
https://github.com/ggevay/malom/pull/3.

Change-Id: I96c93015f9d823cb1a76d4657cf9fffe585ab094