perfect: Fix issue with move actions in Flutter
When using Flutter, there's a high chance that when the perfect AI engine
returns two consecutive moves, 'moving' and 'removing', the frontend
only executes 'removing'. This leads to action execution failure.
Previously, 'removing' was stored and returned directly on the next
search. However, this approach has shown some issues, with the cause
being unclear. The solution is to always return only the first move
from the engine, whether it's two moves or one. This means the action
for 'removing' is recalculated, which might reduce performance but
ensures accuracy. Additionally, when two moves are returned, the result
of 'removing' from the second move might differ from that obtained
through a new search.

Change-Id: I5a17b876c4985a6dbf9447381f1361343263f1d7