Add dynamic description handling in Perfect Database dialog
Modified `use_perfect_database_dialog.dart` to dynamically set the description
based on the `piecesCount` setting. If `piecesCount` is not 9, an experimental
warning is prefixed to the original description.

In `rule_settings_page.dart`, added a condition to show a snackbar message
if the `usePerfectDatabase` setting is enabled and `piecesCount` is changed
to a value other than 9. This informs users that they need to reopen the app
for changes to take effect, ensuring clarity in settings modification.

Change-Id: Icf73b6d0f7ca6b4325b113350da2a4d53b39a712