accessibility: Add semantics to CustomDrawerIcon
Use the Semantics widget to wrap the CustomDrawerIcon and set the
label attribute to S.of(context).menu, indicating that the
function of this icon is to open the drawer.

Through this modification, when the screen reader focuses on this
icon, it will read the label "Menu" to let users using
screen readers know the function of this icon.

But actually the screen reader still reads: "Open navigation menu".