android: Increase compileSdkVersion to 31 for all plugins
This commit updates the `compileSdkVersion` for all plugins to resolve the
`android:attr/lStar not found` error encountered during build. It adjusts the
compile SDK version to 31 to align with AndroidX dependencies and newer Android
SDK requirements. This change is made in the `subprojects` section of the
`android/build.gradle` file to ensure all plugins adhere to this version
without manual updates.

https://github.com/flutter/flutter/issues/153281#issuecomment-2292201697