mini-board: Hide MiniBoard navigation icon when tapping outside
Ensure that the breathing navigation icon on a MiniBoard disappears when tapping
anywhere outside of the MiniBoard. Added a global GestureDetector in
MovesListPage to detect taps and hide the active MiniBoard icon. Modified
MiniBoard's GestureDetector to use HitTestBehavior.opaque to prevent event
propagation. Introduced a static MiniBoardState.hideActiveBoard() method to
allow external calls for hiding the active icon.

Change-Id: I8baff7d7550075e6f58f84938494379cb09e6778