Fix: Append semicolon to Flutter version string on same line
The script previously wrote the terminating semicolon for the Flutter version
map to a new line in the generated $FLUTTER_VERSION_FILE. This resulted
in incorrect Dart syntax or undesirable formatting where the semicolon
appeared as a separate statement.

This commit modifies the script to append the semicolon directly to the end
of the line containing the 'flutter --version --machine' output. This is
achieved by using the 'ex' command to perform an in-place edit of the
$FLUTTER_VERSION_FILE after the 'tee' command has written the version
string. The previous 'echo ";" >> "$FLUTTER_VERSION_FILE"' command, which
caused the newline, has been replaced.

The resulting Dart file will now have the version string correctly
terminated by a semicolon on the same line, for example:
  const Map<String, String> flutterVersion =
  {"channel":"stable","frameworkVersion":"3.x.y",...};

Change-Id: I888bc1efff4de7263f32e938e858f16d69ee8d9e