Added model loading logic to main.py
to ensure the use of the latest best model during self-play and training.

This ensures that the model is loaded when available by checking the model file's existence, and uses the currently initialized model if the model is not found.

This change aims to improve the flexibility and efficiency of model training.

Change-Id: Ia8c44ffe211bf5094c1be3c8904e7315a03c6425