Optimized model loading logic to support smarter checking and training processes.
If the directory does not exist or there is no .tar file → Ignore the load_model setting and start a new training session.

If there is a .tar file but load_model=false → Start a new training session.

If there is a .tar file and load_model=true → Check if the target specific file exists and load it if it exists.

Change-Id: I5d13060c0d159c688268cdd68b4da8beb156ad3b