Optimized logging during the sampling phase
Added calculation of the total number of saved examples and the number of examples in the current iteration.
The end-of-training summary printing logic has also been adjusted.
It now only prints the summary in non-sampling mode and logs relevant information in sampling mode.
This change aims to improve log readability and transparency of the training process.

Change-Id: Ief27510774384dbc6e89fec9eb54b059a8612198