Enhanced debug logging in the PerfectDB class
Detailed debug logging has been added to the evaluate method of the PerfectDB class to record key parameters and status information during the evaluation process.
This includes calculating and recording the number of white and black stones on the board, as well as the expected sector ID.
These changes are intended to improve debugging capabilities and help developers better diagnose issues.

Change-Id: I6c5ad57c12826673528e841ef57a68660bd0f9fd