nnue: Modify parameter types of feature extraction functions to pass position objects by reference
Updated multiple functions in the feature extractor, changing the parameter type of the position object from constant reference to reference to improve performance and consistency. Meanwhile, adjusted the start and end conditions of loops to ensure traversal of all valid board squares. This change helps optimize the feature extraction process.

Change-Id: Ibc9ec99d92e106186e83d7bb6872cf18358e9fa2