nnue: Enhanced Learning Rate Scheduling for Improved Training Efficiency and Stability
Added an adaptive learning rate scheduler to the training scripts, enabling dynamic adjustment of the learning rate based on training and validation losses. Introduced an option for automatic learning rate scaling to compute the initial learning rate based on batch size and dataset size. Additionally, updated the documentation to reflect the new learning rate parameters and scheduling strategies. These improvements aim to enhance the efficiency and stability of model training.

Change-Id: Ibc9ec99d92e106186e83d7bb6872cf18358e9fa2