Updated NNUE Training and Feature Extraction Logic
Main changes:

Modified the database configuration to directly utilize the Perfect DB DLL, eliminating dependency on the engine.
In environment validation, updated the check logic for the engine executable file, making it optional.
Added debug information output during the feature extraction process to facilitate tracking of the feature extraction process and results.
These improvements aim to enhance the flexibility and maintainability of the code while improving the accuracy and debugging capabilities of feature extraction.

Change-Id: I694416def69a11a1f9de986990c4cee260b36a14