Added cumulative time tracking to the training visualizer
This allows for better calculation and display of the total elapsed time and estimated remaining time during training. Updated logging now provides detailed information on the average time per training epoch and the total training time. These improvements are designed to enhance the visualization and user experience of the training process.

Change-Id: I1f41bb48671ac290c388b6bdf472faa0548222d4